refactor: :recycle: simplify public profile component structure and enhance profile sharing functionality
