+++ src/components/public-profile/index.js
@@ -1,0 +1,6 @@
+export { default as ProfileHeader } from './ProfileHeader';
+export { default as StatCard } from './StatCard';
+export { default as AchievementCard } from './AchievementCard';
+export { default as ActivityChart } from './ActivityChart';
+export { default as ShareProfile } from './ShareProfile';
+export { default as ProfileNotFound } from './ProfileNotFound';
